Cleans up ReduxWPAPI settings by decoupling the api binding methods
