+++ src/constants/requestStatus.js
@@ -1,9 +1,3 @@
-export const WAITING = 'WAITING';
-export const READY = 'READY';
-export const ERROR = 'ERROR';
-
-export default {
-  WAITING,
-  READY,
-  ERROR,
-};
+export const pending = 'pending';
+export const resolved = 'resolved';
+export const rejected = 'rejected';
