+++ src/index.js
@@ -1,6 +1,6 @@
 import ReduxWPAPI from './ReduxWPAPI';
 
 export default ReduxWPAPI;
-export * as RequestStatuses from './constants/requestStatus';
+export * as RequestStatus from './constants/requestStatus';
 export { wp } from './actions';
 export { selectQuery } from './selectors';
